requirements.txt
numpy==1.13.3
PyYAML==3.12
redis==3.3.11
git+https://github.com/webpy/webpy#egg=web.py